+++ src/Component/AutoMapper/Tests/Fixtures/ClassWithMapToContextAttribute.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Jane\Component\AutoMapper\Tests\Fixtures;
+
+use Jane\Component\AutoMapper\Attribute\MapToContext;
+
+class ClassWithMapToContextAttribute
+{
+    public function __construct(
+        private string $value,
+    ) {
+    }
+
+    public function getValue(
+        #[MapToContext('prefix')] string $prefix,
+        #[MapToContext('suffix')] string $suffix,
+    ): string {
+        return "{$prefix}_{$this->value}_{$suffix}";
+    }
+}